Rename Report.hpp to Report.h
